separate config from AccountApplication for loose coupling unit test purpose
